refactor: AuthorizationExtractor는 config 대신 auth 패키지로 이동
