+++ backend/src/main/java/com/woowacourse/kkogkkog/auth/presentation/AuthorizationExtractor.java
@@ -1,4 +1,4 @@
-package com.woowacourse.kkogkkog.config;
+package com.woowacourse.kkogkkog.auth.presentation;
 
 import java.util.Enumeration;
 import javax.servlet.http.HttpServletRequest;
